requirements.txt
nibabel
tqdm
numpy
pandas
matplotlib
pyyaml
scikit-image
